More code cleanup
- Reduce `LigeroProver::ligero_prove` to just `ligero_prove` (this
  causes a large diff because lots of code that doesn't change moves
  left by one indent)
- Rename mod `ligero::committer` to `tableau`
- move `TableauLayout` from `ligero` to `ligero::Tableau`
- move `LigeroCommitment` from `ligero::tableau` to `ligero`
- move `ligero::extend` to trait method on
  `LagrangePolynomialFieldElement`.

No actual logic changes.
