+++ src/fields/mod.rs
@@ -219,6 +219,69 @@ pub trait LagrangePolynomialFieldElement: FieldElement {
 
         out
     }
+
+    /// The extend method, as defined in [2.2.1][1] and [2.2.2][2]. We interpolate a polynomial of
+    /// degree at most `nodes.len() - 1` from the provided evaluations at points `[0..nodes.len())`
+    /// and then evaluate that polynomial at `[0, evaluations)`.
+    ///
+    /// The default implementation only works for large characteristic fields.
+    ///
+    /// [1]: https://datatracker.ietf.org/doc/html/draft-google-cfrg-libzk-01#section-2.2.1
+    /// [2]: https://datatracker.ietf.org/doc/html/draft-google-cfrg-libzk-01#section-2.2.2
+    fn extend(nodes: &[Self], evaluations: usize) -> Vec<Self> {
+        // For now we use the relatively straightforward method of computing Lagrange basis polynomials
+        // for each provided point.
+        //
+        // https://en.wikipedia.org/wiki/Lagrange_polynomial#Definition
+        let mut output = Vec::with_capacity(evaluations);
+        let mut denominators = vec![None; nodes.len()];
+
+        for input in (0..evaluations).map(|e| Self::from_u128(e as u128)) {
+            let mut eval = Self::ZERO;
+
+            // Evaluate each basis polynomial
+            for (node_x_coordinate, node_y_coordinate) in nodes.iter().enumerate() {
+                // Compute and cache denominators
+                let denominator = match denominators[node_x_coordinate] {
+                    Some(denominator) => denominator,
+                    None => {
+                        let denominator = nodes
+                            .iter()
+                            .enumerate()
+                            .filter(|(other_node_x_coordinate, _)| {
+                                *other_node_x_coordinate != node_x_coordinate
+                            })
+                            .fold(Self::ONE, |product, (other_node_x_coordinate, _)| {
+                                product
+                                    * (Self::from_u128(node_x_coordinate as u128)
+                                        - Self::from_u128(other_node_x_coordinate as u128))
+                            })
+                            .mul_inv();
+                        denominators[node_x_coordinate] = Some(denominator);
+                        denominator
+                    }
+                };
+
+                // Compute each term of the basis polynomial
+                let basis_poly_eval = nodes
+                    .iter()
+                    .enumerate()
+                    .filter(|(other_node_x_coordinate, _)| {
+                        *other_node_x_coordinate != node_x_coordinate
+                    })
+                    .fold(Self::ONE, |product, (other_node_x_coordinate, _)| {
+                        product * (input - Self::from_u128(other_node_x_coordinate as u128))
+                    })
+                    * *node_y_coordinate
+                    * denominator;
+                eval += basis_poly_eval;
+            }
+
+            output.push(eval);
+        }
+
+        output
+    }
 }
 
 pub fn field_element_iter<FE: CodecFieldElement>() -> impl Iterator<Item = FE> {
@@ -791,5 +854,41 @@ mod tests {
     #[wasm_bindgen_test(unsupported = test)]
     fn lagrange_basis_polynomial_field_2_128() {
         lagrange_basis_polynomial_test::<Field2_128>();
+    }
+
+    fn extend_x_2<FE: LagrangePolynomialFieldElement>() {
+        let output = FE::extend(
+            // x^2 evaluated at 0, 1, 2 => 0, 1, 4
+            &[FE::ZERO, FE::ONE, FE::from_u128(4)],
+            6,
+        );
+
+        assert_eq!(
+            output,
+            // x^2 evaluated at 0..6 = > 0, 1, 4, 9, 16, 25
+            vec![
+                FE::ZERO,
+                FE::ONE,
+                FE::from_u128(4),
+                FE::from_u128(9),
+                FE::from_u128(16),
+                FE::from_u128(25),
+            ]
+        );
+    }
+
+    #[wasm_bindgen_test(unsupported = test)]
+    fn extend_x_2_p128() {
+        extend_x_2::<FieldP128>();
+    }
+
+    #[wasm_bindgen_test(unsupported = test)]
+    fn extend_x_2_p256() {
+        extend_x_2::<FieldP256>();
+    }
+
+    #[wasm_bindgen_test(unsupported = test)]
+    fn extend_x_2_p521() {
+        extend_x_2::<FieldP521>();
     }
 }
